requirements.txt
alembic==1.9.2
bcrypt==4.0.1
bidict==0.22.1
blinker==1.5
cachelib==0.10.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
dnspython==1.16.0
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-Mail==0.9.1
Flask-Migrate==4.0.3
Flask-Session==0.4.0
Flask-SocketIO==5.3.2
Flask-SQLAlchemy==3.0.3
gevent==22.10.2
gevent-websocket==0.10.1
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
loadenv==0.1.1
Mako==1.2.4
MarkupSafe==2.1.2
Pillow==9.4.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
pycparser==2.21
python-dotenv==0.21.1
python-engineio==4.3.4
python-socketio==5.7.2
requests==2.28.2
six==1.16.0
SQLAlchemy==2.0.0
stripe==5.0.0
typing_extensions==4.4.0
urllib3==1.26.14
Werkzeug==2.2.2
zope.event==4.6
zope.interface==5.5.2
